backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
openai==1.93.0
anthropic==0.40.0
langfuse==2.50.3
GitPython==3.1.40
PyYAML==6.0.1
python-dotenv==1.0.0
aiofiles==24.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psutil==5.9.8
